requirements.txt
argparse
PyPDF2
nltk
pydub
moviepy
requests
langchain-ollama
langchain
pydantic
websocket-client
rapidfuzz
tinydb
ebooklib